Moved TimerInterface to a group to be eventually moved into a shared package
