+++ macOS/UnitTests/DefaultBrowserAndAddToDockPrompts/DefaultBrowserAndDockPromptStatusUpdateNotifierTests.swift
@@ -143,45 +143,3 @@ final class DefaultBrowserAndDockPromptStatusUpdateNotifierTests {
         assert(isAddedToDock: true, isDefaultBrowser: true, numberOfEventsReceived: 2)
     }
 }
-
-final class MockTimer: TimerInterface {
-    var isValid: Bool = true
-    private(set) var didCallInvalidate = false
-
-    let timeInterval: TimeInterval
-    let repeats: Bool
-    private let block: (TimerInterface) -> Void
-
-    init(timeInterval: TimeInterval, repeats: Bool, block: @escaping (TimerInterface) -> Void) {
-        self.timeInterval = timeInterval
-        self.repeats = repeats
-        self.block = block
-    }
-
-    func invalidate() {
-        didCallInvalidate = true
-    }
-
-    func fire() {
-        block(self)
-    }
-}
-
-final class MockTimerFactory: TimerCreating {
-    private(set) var didCallMakeTimer = false
-    private(set) var capturedInterval: TimeInterval?
-    private(set) var capturedRepeats: Bool?
-
-    private(set) var createdTimer: MockTimer?
-
-    func makeTimer(withTimeInterval interval: TimeInterval, repeats: Bool, on runLoop: RunLoop, block: @escaping @Sendable (TimerInterface) -> Void) -> TimerInterface {
-        didCallMakeTimer = true
-        capturedInterval = interval
-        capturedRepeats = repeats
-
-        let timer = MockTimer(timeInterval: interval, repeats: repeats, block: block)
-        createdTimer = timer
-        return timer
-    }
-
-}
